requirements.txt
Flask==3.1.2
flask-cors==5.0.0
tensorflow==2.20.0
Pillow==11.0.0
numpy==2.0.2
